@@ -1,15 +1,15 @@
 {
   "name": "pyrrha-js",
-  "version": "1.8.0",
+  "version": "1.8.1-beta1",
   "description": "Pandora Pyrrha functions and utilities library for JavaScript-based applications",
   "private": true,
   "main": "dist/index.js",
   "scripts": {
     "test:build": "docker build --tag pyrrha_js_tests --file ./tests/Dockerfile --build-arg PYRRHA_CONSENSUS_BRANCH=develop .",
     "pretest": "npm run build && npm run test:build",
-    "pretest-with-coverage": "npm run test:build",
-    "test": "docker run -e WITH_COVERAGE=0 -it pyrrha_js_tests",
-    "test-with-coverage": "docker run -e WITH_COVERAGE=1 -it pyrrha_js_tests",
+    "pretest-with-coverage": "npm run --rm test:build",
+    "test": "docker run --rm -e WITH_COVERAGE=0 -it pyrrha_js_tests",
+    "test-with-coverage": "docker run --rm -e WITH_COVERAGE=1 -it pyrrha_js_tests",
     "build": "npx rimraf ./dist && NODE_ENV=production npx babel ./src --out-dir ./dist && npx cpy ./package.json ./dist"
   },
   "repository": {
@@ -21,19 +21,19 @@
   },
   "homepage": "https://github.com/pandoraboxchain/pyrrha-js",
   "dependencies": {
-    "ipfs-api": "^24.0.1",
-    "web3": "1.0.0-beta.34"
+    "ipfs-api": "25.0.0",
+    "web3": "1.0.0-beta.36"
   },
   "devDependencies": {
-    "@babel/cli": "7.0.0",
-    "@babel/core": "7.0.0",
+    "@babel/cli": "7.1.2",
+    "@babel/core": "7.1.2",
     "@babel/plugin-proposal-object-rest-spread": "7.0.0",
     "@babel/plugin-syntax-object-rest-spread": "7.0.0",
     "@babel/polyfill": "7.0.0",
-    "@babel/preset-env": "7.0.0",
+    "@babel/preset-env": "7.1.0",
     "@babel/register": "7.0.0",
     "cpy-cli": "2.0.0",
-    "eslint": "5.4.0",
+    "eslint": "5.7.0",
     "rimraf": "2.6.2",
     "github-utils": "git+https://github.com/pandoraboxchain/github-utils.git#v0.1.4"
   },
