@@ -14,6 +14,7 @@ export const WRONG_TYPE = 'WRONG_TYPE';
 export const ADDRESS_REQUIRED = 'ADDRESS_REQUIRED';
 export const WEB3_REQUIRED = 'WEB3_REQUIRED';
 export const WEB3_NOT_CONNECTED = 'WEB3_NOT_CONNECTED';
+export const WEB3_CONNECTION_TIMEOUT = 'WEB3_CONNECTION_TIMEOUT';
 export const CONTRACT_REQUIRED = 'CONTRACT_REQUIRED';
 export const IPFS_REQUIRED = 'IPFS_REQUIRED';
 export const IPFS_NOT_CONNECTED = 'IPFS_NOT_CONNECTED';
@@ -22,8 +23,9 @@ export const TRANSACTION_UNSUCCESSFUL = 'TRANSACTION_UNSUCCESSFUL';
 export const FAILURE_EVENT = 'FAILURE_EVENT';
 export const EXPECT_NOT_A_MEMBER = 'EXPECT_NOT_A_MEMBER';
 export const EXPECT_TYPE_OPTIONS_REQUIRED = 'EXPECT_TYPE_OPTIONS_REQUIRED';
+export const PJS_REQUIRED = 'PJS_REQUIRED';
 
-export default (code, key, type, value, ...args) => {
+export default (code, ...args) => {
     let message = 'Unknown error';
         
     switch (code) {
@@ -32,7 +34,7 @@ export default (code, key, type, value, ...args) => {
             break;
 
         case WRONG_TYPE:
-            message = `Wrong property type. Expected '${key}' to be: '${type}', but got: '${value}'`;
+            message = `Wrong property type. Expected '${args[0]}' to be: '${args[1]}', but got: '${args[2]}'`;
             break;
 
         case WEB3_REQUIRED:
@@ -41,6 +43,10 @@ export default (code, key, type, value, ...args) => {
 
         case WEB3_NOT_CONNECTED:
             message = 'Web3 not connected to provider';
+            break;
+
+        case WEB3_CONNECTION_TIMEOUT:
+            message = `Websocket connection timeout (${args[0]}ms) exceeded`;
             break;
 
         case WEB3_METAMASK_REQUIRED:
@@ -52,7 +58,7 @@ export default (code, key, type, value, ...args) => {
             break;
 
         case ADDRESS_REQUIRED:
-            message = `Address of '${args[0] || key}' is required. Wrong ethereum address: ${value}`;
+            message = `Address of '${args[3] || args[0]}' is required. Wrong ethereum address: ${args[2]}`;
             break;
 
         case IPFS_REQUIRED:
